requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
statsmodels==0.14.2
prophet==1.1.5
ta==0.11.0
ccxt==4.3.84
joblib==1.4.2
fastapi==0.114.0
uvicorn==0.30.6
optuna==3.6.1
torch==2.3.1 ; platform_system != "Darwin" or platform_machine != "arm64"
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
